Multithreading Configuration and Failure Handling for markBestSubGraphsInTime(int initialDelay, int interval)
the explanation lines that possibly blocked multithreading mode was eliminated in application.conf. The timeout period for markBestSubGraphsInTime(int initialDelay, int interval) method was increased to 180 from 60 and the exception handling mechanism in the same method started to trigger the same method in case of exceptions. The initial null values of each benchmark in the same method was changed to -1. mark Subgraph methods in AlignerImpll are modified to send -1 values when they throw an exception.