removeBadMappingsWhenUnimproved method implemented and removeBadMappingsToReduceInduction1 method has been revised
The functionality of removeBadMappingsWhenUnimproved implemented in the AlignerImpl class and its execution is carried out implicitly AlignerImpl class instead of the addRandomMapping method in AkkaSystem class so that it is more tightly coupled with the parameters that it requires. 
the functionality of compareSimilarity method is revised to only let the removal of weaker mappings with respect to certain treshold values by primarily considering a proper combination of the benchmarks. The changes aimed to effect the removeBadMappingsToReduceInduction1 which was also revised to compute the exact number of removed mappings.
New contents has been added to the addRandomMapping method in order to improve results.